Removed dead code, fixed few column type bugs.

+++ src/db/models/Quiz.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, ForeignKey, Boolean
+from sqlalchemy import Column, Integer, String, ForeignKey, Boolean, BigInteger
 from sqlalchemy.orm import relationship, backref
 from sqlalchemy.ext.hybrid import hybrid_property
 
@@ -36,7 +36,7 @@ class Quiz(BaseModel):
 
     id = Column(Integer, primary_key=True)
     name = Column(String(50), nullable=False)
-    author_id = Column(Integer, ForeignKey(User.id, ondelete='CASCADE'), nullable=False)
+    author_id = Column(BigInteger, ForeignKey(User.id, ondelete='CASCADE'), nullable=False)
     is_available = Column(Boolean, default=True, nullable=False)
     is_public = Column(Boolean, default=True, nullable=False)
     is_statistical = Column(Boolean, default=False, nullable=False)
